refactor: add profiling, simplify search code
