+++ lib/constraint.ts
