+++ lib/matrix.ts
